% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxaSPU.R
\name{approx_aSPU}
\alias{approx_aSPU}
\title{The function for adaptive sum of powered score (aSPU) test}
\usage{
approx_aSPU(betas, vcov_mat, pow = c(1:8, Inf), B)
}
\arguments{
\item{betas}{the vector (length = p) of maximum likelihood estimates; p is the number of paramters to be tested.}

\item{vcov_mat}{the p by p variance-covariance matrix of MLE.}

\item{pow}{the power vector for aSPU tests.}

\item{B}{number of simulations}
}
\value{
simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test
}
\description{
It gives the simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test.
}
\details{
the suggested power term is c(1:8, Inf)
}
\examples{
# install.packages(lme4)
library(lme4)
library(MASS)
data(exdat_GLMM)
fit1<-glmer(Y~.-ID + (1|ID), ,family="binomial",data=dat,control=glmerControl(optimizer="bobyqa"))
betas<-fixef(fit1)[-1]
vcov_mat<-vcov(fit1)[-1,-1]
approx_aSPU(betas=betas,vcov_mat = vcov_mat,pow=c(1:8,Inf),B=1000) 
}
\author{
Zhiyuan (Jason) Xu, Yiwei Zhang and Wei Pan
}
\references{
Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014) A powerful and adaptive
association test for rare variants, Genetics, 197(4), 1081-95

Yiwei Zhang, Zhiyuan Xu, Xiaotong Shen, Wei Pan (2014) Testing for association with multiple
traits in generalized estimation equations, with application to neuroimaging data. Neuroimage.
96:309-25
}

