\name{approxaSPU-package}
\alias{approxaSPU-package}
\alias{approxaSPU}
\docType{package}
\title{
\packageTitle{approxaSPU}
}
\description{
\packageDescription{approxaSPU}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{approxaSPU}
\packageIndices{approxaSPU}
how to use the package approx_aSPU()
}
\author{
\packageAuthor{approxaSPU}

Maintainer: \packageMaintainer{approxaSPU}
}
\references{
Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014) A powerful and adaptive
association test for rare variants, Genetics, 197(4), 1081-95

Yiwei Zhang, Zhiyuan Xu, Xiaotong Shen, Wei Pan (2014) Testing for association with multiple
traits in generalized estimation equations, with application to neuroimaging data. Neuroimage. 96:309-25
}


\keyword{ approximate score, aSPU }
\seealso{

\code{\link[aSPU:<aSPU>-package]{<aSPU>}}
}

